+++ dsr_agents/src/agent_node.cpp
@@ -11,7 +11,7 @@
 
 // ROS
 #include "nav2_util/node_utils.hpp"
-#include "tf2_geometry_msgs/tf2_geometry_msgs.h"
+#include "tf2_geometry_msgs/tf2_geometry_msgs.hpp"
 
 #include "dsr_agents/agent_node.hpp"