+++ dsr_agents/src/generic_agent.cpp
@@ -23,7 +23,6 @@
 
 // DSR
 #include "dsr_agents/qt_executor.hpp"
-#include "dsr_agents/ros_to_dsr_types.hpp"
 #include "dsr_agents/generic_agent.hpp"
 
 /* Initialize the publishers and subscribers */